:recycle: Refactor: Chatting & Message 관련 코드 제거
