--- src/main/java/briefing/chatting/domain/Message.java
@@ -1,53 +1,0 @@
-package briefing.chatting.domain;
-
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import lombok.*;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Message {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Chatting chatting;
-
-    @Enumerated(EnumType.STRING)
-    private MessageRole role;
-
-    @Column(nullable = false, length = 1024)
-    private String content;
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    public Message(final Chatting chatting, final MessageRole role, final String content) {
-        this.chatting = chatting;
-        this.role = role;
-        this.content = content;
-    }
-}
