--- src/main/java/briefing/chatting/domain/repository/ChattingRepository.java
@@ -1,9 +1,0 @@
-package briefing.chatting.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import briefing.chatting.domain.Chatting;
-
-@Repository
-public interface ChattingRepository extends JpaRepository<Chatting, Long> {}
